+++ packages/services/builder/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @0xsequence/builder
 
+## 3.0.0-beta.5
+
+### Patch Changes
+
+- Beta release for v3
+
+## 3.0.0-beta.4
+
+### Patch Changes
+
+- RC5 upgrade
+
+## 3.0.0-beta.3
+
+### Patch Changes
+
+- 3.0.0-beta.3 with fixes
+
+## 3.0.0-beta.2
+
+### Patch Changes
+
+- 3.0.0-beta.2 with identity instrument updates
+
+## 3.0.0-beta.1
+
+### Patch Changes
+
+- 3.0.0-beta.1
+
 ## 2.3.8
 
 ### Patch Changes